requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
jiwer>=3.0.0
nltk>=3.8.1
rouge-score>=0.1.2
tqdm>=4.65.0
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1
